requirements.txt
Django~=3.0.11
djangorestframework
djangogrpcframework
grpcio
grpcio-tools